requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.6.0
requests==2.31.0
certifi==2024.2.2
urllib3==2.2.0